Add muliselect flavor form
Ok, I did get this working. But it is not really nice. In the standard HTML version,  you need to keep CTRL or SUPER (on mac) pressed to select multiple items. If you click a single element then it resets the filed to only that value.

When operating with react, I only get a single value with the even. This makes it a bit easier to select multiple values, because you do not have to keep an additional button pressed for multi select. The weird thing is that on every click, the selection goes completely away and re-renders. While this works correctly, this looks odd.

I am also realizing again, the select if not a good interface for selecting multiple values. I think checkboxes should be more appropriate.

Also, Gatsby/ESLint is throwing warnings for using `onChange` with the `multiple` options.