Add non-editable input fields
There are two ways to disable input fields. One is to define a `value` but not
an `onChange` handler. This prevents the user from editing the field, but the
field remains looking normal.

This lock onlu works when the predefined value is different from `null` or `undefined`.

Because the visuals still suggest an editable field, it might be confusing to users
when the field is actually not editable. To visually disable a field, use the
`disabled` flag attribute. This disables user input to the field, no matter what
the value is.