Use onBlur instead of onChange on select
Gatsby was showing warnings that `onChange` could cause issues on screen readers.
The proposed fix was to use `onBlur` instead. That is what I switched to.
This
switch caused another warning, that said to use `defaultValue` instead of `value`
to define the default value. That is also implemented.

Still works as expected, but without the warnings.