Create separate component to create fields
Before, the field elements where created in the form component. This
was fine as there was not case handling, labels or error display
elements. But, all the aforementioned is going to be added soon. So the
first step was to create the simple input fields in a separate
component.

This was already valuable as I found out that the `ref` attribute can
not be simply passed to other components. The child component has to
be created with `React.forwardRef` so that the special `ref` attribute
can be passed from one component to another.

https://reactjs.org/docs/forwarding-refs.html