Add checkbox field case
This is only for the single checkbox. For the checkbox, the block
display label does not make much sense. It is much more common to use
the on the same line as the checkbox.

The next case is multiple checkboxes. I think these two cases need to
be balanced some how. Therefore, I have not yet changed the styling of
the label for checkboxes.

Also, I had wrongly thought that the inputs `name` attribute is used to
connect the label to the input, but it is the `id`. This is fixed now.
This creates a larger click target, which is especially useful for
checkboxes. It is also helpful for other field types.