Create generic field type
Most field types where nearly identical, except for the `type` attribute
that was selected. To reduce the amount of code required, a
`GenericField` component has been added that passes the `type` argument
through to the input element.

This is saving a tremendous amount of code and makes the logic of the
wrapping `Field` component much cleaner.