Move label styling from from to field
Handling this on the form level is going to be dirty. The placement of
the label is depended on the field type. It's above for most fields,
but the checkbox needs the label next to it.

Since there already is a special component for the field, this component
is now also handling the creation and placement of the label.