Add display element for multiple checkboxes
This is only the logic to display the field so far. Extracting multiple
values seems to be a bit tricker than it has been sofar.

This is because I currently attach the reference to the field. For all
elements so far, meant that one field from Wagtail is represented with
one field in the HTML and JS. This is not the case with the
multi-checkboxes. Now I need to create an input for each choice that
is mentioned in the `choices` string. Also, I need to keep separate
references to each of these fields. This is necessary so that they are
not using all the same reference. This leads to only the last choice
being actually available.

So I need to attach the references to something other than the field,
because I need multiple.