Implement value extraction for multi checkboxes
This required a different solution than the previously implemented
fields. The reason for this is that for the previous fields, one field
in Wagtail mapped to one field in the DOM. This is not the case for
multiple checkboxes. Here I need one input field in the DOM for each
choice that is listed in Wagtail.