Use same `name` attribute for multi checkboxes
Although it worked previously with using the checkbox value for the checkbox name,
this is not how the name field in a multi checkbox field is usually used.

To use it correctly as in a normal HTML form, the `name` attribute for all related
checkboxes should be the same. This name is now set to the Wagtail field name,
rather than the value of the individual field.