Add value extraction for drop down field
Before, I forget to add the `ref` attribute to the `select` element.
This caused the value extraction to fail.

The `ref` attribute is added now and the dropdown value can correctly
be extracted.

Also, I moved the data extraction into a separate method from the
submission handler. This is to keep these concerns cleanly separated.