Add datalist element for input with choices
HTML 5 provides the `datalist` element. This element can contain `option` elements
and can be associated with any `input` field. The exact effect of using the
`datalist` element depends on the OS and the browser, but in gernal, the options
will provide suggested values while the user is typing.

This can help to prevent duplicate entries due to capitalization or misspelling.

The implementation adds a datalist element for `input` fields, when choices are
provided.