Move creation of defaultValue array to form
There are multiple field types that take multiple default values.
Before, the splitting of the defaults was handled in the different field
components. This meant repetitive code and actually made the components
logic dirtier.

Now, the splitting of the default string value into an array is handled
in the form once. This makes the handling in the components a lot
cleaner.