Add handling for radio field type
Because of the strong similarity in the display, this is unified with
the multiple checkbox field type.

The data extraction has to differ between radio and checkboxes. A radio
field has multiple options, but only allows a single value to be
selected. In case somebody should mess with the form and select multiple
radio buttons, the data extraction goes over all possible choices and
only retains the last as the selected field.