Limit creation of datalist element to necessary
Before, the datalist element was created on every input element. The
list was empty if no choices where defined. This was the case for most
fields.

What was unexpected about this behavior is that the datalist element
should not be created at all if it does not contain any child elements/
options.

The bug was caused by a check that does not work in JS. `[] === []` is
false in JS... To check if an array is empty, use `Array.length === 0`.

This is fixed now. The datalist element is only created when it contains
options.