Add date input
This was not quite as straight forward. Because of the different browser
support for the HTML native `date` input, I chose to create an input
widget with a third party library `react-datepicker`.

The date is extracted as a JS `Date` object. This allows me to format
the data into what ever I need on the backend.