Add basic setup to handle form as state machine
Handling the form as a state machine is attractive, because this should
make it simple to implement the typical work flow of a form.

Right now, only the skeleton for this handling has been added. But, the
full functionality with 4 states (initial, submitting, success and
error) and 3 actions/events (submit, reject, accept) can be added in
this fashion.

The basic concept is to create custom events for the distinct actions.
Each of these actions will trigger the transition function. The
transition function/method can then determine the next state to move to
based on the current state and the type of action/event.

For a simple form flow, I have identified the following state
transitions:

Initial -> submit -> Submitting
Submitting -> accept -> Success
Submitting -> reject -> Error
Error -> submit -> Submitting

This transition "table" can be extended to allow for more complex logic
flows.