Add first state transition initial -> submitting
This is the first simple state transition. This is only to figure out
the mechanics and has no influence on the displayed form yet.

The transition detection has to be done with nested if-statements
because switch statements do not work,
because  `["a", "b"] === ["a", "b"]` is `false`. Freaking JS again...