Simplify transition function
Before I used nested if statements and had a bunch of console logs in there.
That was good for debugging. Now that the code is working, I can remove
these statements and consolidated the nested if statements into a single statement
with a connected condition. This is easier to read.